Moved all schemas to connect.schemas
