+++ connect/models/__init__.py
@@ -4,29 +4,22 @@
 # Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 
 from .activation_response import ActivationTemplateResponse, ActivationTileResponse
-from .asset import Asset, AssetSchema
-from .base import BaseModel, BaseSchema
-from .company import Company, CompanySchema
-from .connection import Connection, ConnectionSchema
-from .contact import Contact, ContactInfo, ContactInfoSchema, ContactSchema, \
-    PhoneNumber, PhoneNumberSchema
-from .event import EventInfo, EventInfoSchema, Events, EventsSchema
-from .fulfillment import Fulfillment, FulfillmentSchema
-from .hub import Hub, HubInstance, HubInstanceSchema, ExtIdHub, HubSchema, ExtIdHubSchema, \
-    HubStats, HubStatsSchema
-from .marketplace import Activation, ActivationSchema, Agreement, AgreementSchema, AgreementStats,\
-    AgreementStatsSchema, Contract, ContractSchema, Marketplace, MarketplaceSchema
-from .parameters import Constraints, ConstraintsSchema, Param, ParamSchema, ValueChoice, \
-    ValueChoiceSchema
-from .product import CustomerUiSettings, CustomerUiSettingsSchema, Document, DocumentSchema, \
-    DownloadLink, DownloadLinkSchema, Item, ItemSchema, Product, ProductConfiguration, \
-    ProductConfigurationSchema, ProductSchema, Renewal, RenewalSchema
-from .server_error_response import ServerErrorResponse, ServerErrorResponseSchema
-from .tier_config import Account, AccountSchema, Template, TemplateSchema, TierConfig, \
-    TierConfigRequest, TierConfigRequestSchema, TierConfigSchema
-from .tiers import Tier, Tiers, TierSchema, TiersSchema
-from .usage import UsageFile, UsageFileSchema, UsageListing, UsageListingSchema, UsageRecord, \
-    UsageRecords, UsageRecordSchema, UsageRecordsSchema
+from .asset import Asset
+from .base import BaseModel
+from .company import Company
+from .connection import Connection
+from .contact import Contact, ContactInfo, PhoneNumber
+from .event import EventInfo, Events
+from .fulfillment import Fulfillment
+from .hub import Hub, HubInstance, ExtIdHub, HubStats
+from .marketplace import Activation, Agreement, AgreementStats, Contract, Marketplace
+from .parameters import Constraints, Param, ValueChoice
+from .product import CustomerUiSettings, Document, DownloadLink, Item, Product, \
+    ProductConfiguration, Renewal
+from .server_error_response import ServerErrorResponse
+from .tier_config import Account, Template, TierConfig, TierConfigRequest
+from .tiers import Tier, Tiers
+from .usage import UsageFile, UsageListing, UsageRecord, UsageRecords
 
 __all__ = [
     'Account',
@@ -71,46 +64,4 @@ __all__ = [
     'UsageRecord',
     'UsageRecords',
     'ValueChoice',
-
-    # Schemas
-    'AccountSchema',
-    'ActivationSchema',
-    'AgreementSchema',
-    'AgreementStatsSchema',
-    'AssetSchema',
-    'BaseSchema',
-    'CompanySchema',
-    'ConnectionSchema',
-    'ConstraintsSchema',
-    'ContactInfoSchema',
-    'ContactSchema',
-    'ContractSchema',
-    'CustomerUiSettingsSchema',
-    'DocumentSchema',
-    'DownloadLinkSchema',
-    'EventInfoSchema',
-    'EventsSchema',
-    'ExtIdHubSchema',
-    'FulfillmentSchema',
-    'HubInstanceSchema',
-    'HubSchema',
-    'HubStatsSchema',
-    'ItemSchema',
-    'MarketplaceSchema',
-    'ParamSchema',
-    'PhoneNumberSchema',
-    'ProductConfigurationSchema',
-    'ProductSchema',
-    'RenewalSchema',
-    'ServerErrorResponseSchema',
-    'TemplateSchema',
-    'TierConfigRequestSchema',
-    'TierConfigSchema',
-    'TierSchema',
-    'TiersSchema',
-    'UsageFileSchema',
-    'UsageListingSchema',
-    'UsageRecordSchema',
-    'UsageRecordsSchema',
-    'ValueChoiceSchema',
 ]
