+++ connect/models/company.py
@@ -3,17 +3,8 @@
 # This file is part of the Ingram Micro Cloud Blue Connect SDK.
 # Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 
-from marshmallow import fields, post_load
-
-from .base import BaseModel, BaseSchema
-
-
-class CompanySchema(BaseSchema):
-    name = fields.Str()
-
-    @post_load
-    def make_object(self, data):
-        return Company(**data)
+from .base import BaseModel
+from connect.models.schemas import CompanySchema
 
 
 class Company(BaseModel):
