+++ connect/models/hub.py
@@ -3,51 +3,12 @@
 # This file is part of the Ingram Micro Cloud Blue Connect SDK.
 # Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 
-from marshmallow import Schema, fields, post_load
 from typing import Optional
 
-from .base import BaseModel, BaseSchema
-from .company import Company, CompanySchema
-from .event import Events, EventsSchema
-
-
-class HubInstanceSchema(BaseSchema):
-    type = fields.Str()
-
-    @post_load
-    def make_object(self, data):
-        return HubInstance(**data)
-
-
-class HubStatsSchema(BaseSchema):
-    connections = fields.Int()
-    marketplaces = fields.Int()
-
-    @post_load
-    def make_object(self, data):
-        return HubStats(**data)
-
-
-class HubSchema(BaseSchema):
-    name = fields.Str()
-    company = fields.Nested(CompanySchema)
-    description = fields.Str(allow_none=True)
-    instance = fields.Nested(HubInstanceSchema)
-    events = fields.Nested(EventsSchema)
-    stats = fields.Nested(HubStatsSchema)
-
-    @post_load
-    def make_object(self, data):
-        return Hub(**data)
-
-
-class ExtIdHubSchema(Schema):
-    hub = fields.Nested(HubSchema, only=('id', 'name'))
-    external_id = fields.Str()
-
-    @post_load
-    def make_object(self, data):
-        return ExtIdHub(**data)
+from .base import BaseModel
+from .company import Company
+from .event import Events
+from connect.models.schemas import HubInstanceSchema, HubStatsSchema, HubSchema, ExtIdHubSchema
 
 
 class HubInstance(BaseModel):
