+++ connect/models/server_error_response.py
@@ -3,20 +3,10 @@
 # This file is part of the Ingram Micro Cloud Blue Connect SDK.
 # Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 
-from marshmallow import Schema, fields, post_load
 from typing import List
 
 from .base import BaseModel
-
-
-class ServerErrorResponseSchema(Schema):
-    error_code = fields.Str()
-    params = fields.Dict()
-    errors = fields.List(fields.Str())
-
-    @post_load
-    def make_object(self, data):
-        return ServerErrorResponse(**data)
+from connect.models.schemas import ServerErrorResponseSchema
 
 
 class ServerErrorResponse(BaseModel):
